Create separate client component for moving car
